Split ArenaPopup into a create and edit variant
