+++ src/com/company/popup/ArenaPopup.java
@@ -5,8 +5,8 @@ import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.HBox;
 
-public class ArenaPopup extends Popup<Arena> {
-    private TextField textField;
+public abstract class ArenaPopup extends Popup<Arena> {
+    protected TextField textField;
 
     public ArenaPopup(Arena item) {
         super(item);
@@ -14,25 +14,11 @@ public class ArenaPopup extends Popup<Arena> {
         Label label = new Label("Arena name:");
         textField = new TextField();
         setInputArea(new HBox(label, textField));
-
-        if (item == null)
-            return;
-        textField.setText(item.getArenaName());
     }
 
     @Override
-    public void apply() {
-        if (item == null) {
-            new Arena(textField.getText());
-        } else {
-            item.setArenaName(textField.getText());
-        }
-    }
+    public abstract void apply();
 
     @Override
-    public void delete() {
-        if (item != null) {
-            item.remove();
-        }
-    }
+    public abstract void delete();
 }
